+++ src/main/java/com/unicalsocial/backend/message/ConversationService.java
@@ -1,5 +1,5 @@
 package com.unicalsocial.backend.message;
 
 public interface ConversationService {
-    Integer findByParticipants(Long participant1Id, Long participant2Id);
+    Integer findByParticipants(Integer participant1Id, Integer participant2Id);
 }
